feat: add widget card and code block components for testimonials
- Implemented `CodeBlock` component for displaying and copying code snippets.
- Created `WidgetCard` component to showcase testimonials with a preview dialog and code generation.
- Refactored `WidgetsPage` to integrate new components and manage API key section.
- Removed unused code related to API key management and documentation.
